minor refactor and documentation
